requirements.txt
-i https://pypi.org/simple
certifi==2019.11.28
chardet==3.0.4
click==7.0
flask==1.1.1
future==0.18.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
jinja2==2.11.1
line-bot-sdk==1.16.0
markupsafe==1.1.1
requests==2.23.0
urllib3==1.25.8
werkzeug==1.0.0
